==> _goalkeeper_analysis/_goalies_analysis/final_work/1-model_expected_end_cost.ipynb <==
## Regression to see the relationship between the features and the player's "end_cost"
"end_cost" is the cost of a player at the end of each fantasy season. We can use it to see how the fanstay algo values a player's impact over the course of the season

import pandas as pd
import numpy as np
import requests
from sklearn.model_selection import train_test_split
from sklearn.metrics import mean_absolute_error, mean_squared_error, r2_score
from sklearn.linear_model import LinearRegression
from matplotlib import pyplot as plt
from sklearn.ensemble import RandomForestRegressor
from sklearn.preprocessing import StandardScaler


#Import goalie-specific, model ready csv

goalie_hist_model_ready = pd.read_csv("C:/Users/Daniel Quinn/Desktop/Bootcamp/Project_2/data/processed/goalie_hist_model_ready.csv")
#Ensure we can read the whole dataframe, without "..."
pd.set_option("display.max_columns", None)
goalie_hist_model_ready = goalie_hist_model_ready.drop(columns = ['Unnamed: 0'])
goalie_hist_model_ready.shape


# Create saves percentage stat - will use NaN's to drop irrelevant goalies

#Create Saves Percentage stat:

goalie_hist_model_ready["Saves_Percentage"] = (
    (goalie_hist_model_ready['saves'] + goalie_hist_model_ready['penalties_saved']) /
    (goalie_hist_model_ready['saves'] + goalie_hist_model_ready['penalties_saved'] + goalie_hist_model_ready['goals_conceded'])
) * 100
goalie_hist_model_1_ready = goalie_hist_model_ready.dropna()
goalie_hist_model_1_ready.shape

# P-Value test on statsmodel regression

I ran a number of different models here. Starting on with a maximal features list, I used each inteatio to pare back the stat. insignificant.

Foir example, "start_cost" overwhelmed the model initially, so I removed it and kept doing so until I got a model that returned p values that made sense.

What "made sense?" Intuitively, it made no sense for a "start_cost" to essentially ignore saves and clean sheets when assessing the value of a goalie at the end of the season.



#Check the p-value to determine the statistical significance of each features

goalie_hist_model_2_ready = goalie_hist_model_1_ready[['minutes', 'total_points', 'clean_sheets', 'goals_conceded',
                                                       'saves', 'ict_index', 'bonus', 'bps',
                                                       'expected_goals_conceded', 'starts', 'penalties_saved', 'expected_goal_involvements',
                                                       'assists', 'own_goals', 'end_cost']]

import statsmodels.api as sm

#create X & y variables
X = goalie_hist_model_2_ready.drop(columns = ['end_cost'])
y = goalie_hist_model_2_ready['end_cost']

#test-training split
X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=42)

# Use the statsmodels package to create and fit a linear regression
lr = sm.OLS(y_train, X_train).fit()
lr.pvalues.sort_values(ascending=False)

# Linear Regression:

#create X & y variables
X = goalie_hist_model_2_ready.drop(columns = ['end_cost'])
y = goalie_hist_model_2_ready['end_cost']

#test-training split
X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=42)

#create model
model = LinearRegression()

model.fit(X_train, y_train)
print("Train model score: ", model.score(X_train, y_train))
print("Test model score: ", model.score(X_test, y_test))

#Make predictions

prediction1 = model.predict(X_test)

#Evaluate models with mse and r2

mse = mean_squared_error(y_test, prediction1) # how close are the predicted values to actual values via the squared differences between expected and real
r2 = r2_score(y_test, prediction1) # r2 - how well do the indep variables explain the variation in the dep var? ) 0 is a perfect model, the larger the nuber, the worse the model is performing

print(f"All Features (end_cost = y):")
print(f"mean squared error (MSE): {mse}")
print(f"R-squared (R2): {r2}")

# LR Model for new data:

#Run the model to get the predicted end costs for each goalie

predicted_end_cost = goalie_hist_model_2_ready.drop(columns = ['end_cost'])
y_new_pred = model.predict(predicted_end_cost)
y_new_pred
exp_end_cost_1 = pd.DataFrame(y_new_pred)
exp_end_cost_1

# Random Forest:

#create X & y variables
X = goalie_hist_model_2_ready.drop(columns=['end_cost'])
y = goalie_hist_model_2_ready['end_cost']

#test-training split
X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=42)

#create & train model
random_forest = RandomForestRegressor(n_estimators=500, random_state=42).fit(X_train, y_train)

# Evaluate the model
print(f"Training Score: {random_forest.score(X_train, y_train)}")
print(f"Testing Score: {random_forest.score(X_test, y_test)}")

predicted_end_cost_1 = goalie_hist_model_2_ready.drop(columns = ['end_cost'])
y_new_pred_1 = random_forest.predict(predicted_end_cost_1)
y_new_pred_1
exp_end_cost_2 = pd.DataFrame(y_new_pred_1)
exp_end_cost_2

# Feature Importance
feature_importances = random_forest.feature_importances_

feature_importances_df = pd.DataFrame(feature_importances, X.columns)


print(feature_importances_df.sort_values(by=0, ascending=False))

# Running models on a different combination of features

# These are using an edited features list - everything here is goalie-centric
#  & not necessarily applicable to other positions - what we used for the p-value test above

#New feature set

edit2_goalies_list = goalie_hist_model_2_ready
# [['expected_assists', 'starts', 'ict_index',
#                                               'red_cards', 'expected_goals_conceded', 'expected_goals',
#                                               'bps', 'expected_goal_involvements', 'yellow_cards',
#                                               'minutes', 'own_goals', 'goals_scored', 'clean_sheets', 'assists',
#                                               'penalties_saved', 'bonus', 'total_points', 'saves', 'goals_conceded', 'end_cost']]

edit2_goalies_list.columns


# Pvalues below are about the same as above:

#Check the p-value to determine the statistical significance of each features

goalie_hist_model_3_ready = edit2_goalies_list[['minutes', 'total_points', 'clean_sheets', 'goals_conceded',
                                                       'saves', 'ict_index', 'bonus', 'bps',
                                                       'expected_goals_conceded', 'starts', 'penalties_saved', 'expected_goal_involvements',
                                                       'assists', 'own_goals', 'end_cost']]

import statsmodels.api as sm

#create X & y variables
X = goalie_hist_model_3_ready.drop(columns = ['end_cost'])
y = goalie_hist_model_3_ready['end_cost']

#test-training split
X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=42)

# Use the statsmodels package to create and fit a linear regression
lr = sm.OLS(y_train, X_train).fit()
lr.pvalues.sort_values(ascending=False)

# Linear Regression:

#create X & y variables
X = edit2_goalies_list.drop(columns=['end_cost'])
y = edit2_goalies_list['end_cost']

#test-training split
X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=42)

#create model
model = LinearRegression()

model.fit(X_train, y_train)
print("Train model score: ", model.score(X_train, y_train))
print("Test model score: ", model.score(X_test, y_test))

#These are slightly lower than above

#Make predictions

prediction1 = model.predict(X_test)

#Evaluate models with mse and r2

mse = mean_squared_error(y_test, prediction1)
r2 = r2_score(y_test, prediction1)

#Evaluate & print the model scores
print(f"All Features (end_cost=y):")
print(f"mean squared error (MSE): {mse}")
print(f"Test R-squared (R2): {r2}")

# without start_cost, the MSE is 10.75; r2 = 0.63

#Full-feature: MSE 1.986 R2 .944

#Random Forest
#create X & y variables
X = edit2_goalies_list.drop(columns=['end_cost'])
y = edit2_goalies_list['end_cost']

#test-training split
X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=42)

#create & train model
random_forest_2 = RandomForestRegressor(n_estimators=500, random_state=42).fit(X_train, y_train)

# Evaluate the model
print(f"Training Score: {random_forest.score(X_train, y_train)}")
print(f"Testing Score: {random_forest.score(X_test, y_test)}")


# Feature Importance
feature_importances = random_forest.feature_importances_

feature_importances_df = pd.DataFrame(feature_importances, X.columns)


print(feature_importances_df.sort_values(by=0, ascending=False))



#The total_points metric is a fantasy-generated metric that includes minutes, clean_sheets, saves,
# penalties_saved, own_goals, goals_conceded. 

# New dataset predictions:

#Run the model to get the predicted end costs for each goalie

predicted_end_cost = edit2_goalies_list.drop(columns = ['end_cost'])
y_new_pred2 = random_forest_2.predict(predicted_end_cost)
print(y_new_pred2)

# Import data & create dataframe to hold y_new_pred2 prediction

y_new_pred2.shape

# Create new dataframe with expected end_costs

end_cost_predictions_df_master = goalie_hist_model_1_ready
end_cost_predictions_df_master['Predicted_End_Cost'] = y_new_pred2
end_cost_predictions_df_master = end_cost_predictions_df_master[['id', 'first_name', 'second_name', 'team', 'element_type', 'code',
       'element_code', 'season', 'total_points',
       'minutes', 'goals_scored', 'assists', 'clean_sheets', 'goals_conceded',
       'own_goals', 'penalties_saved', 'penalties_missed', 'yellow_cards',
       'red_cards', 'saves', 'bonus', 'bps', 'influence', 'creativity',
       'threat', 'ict_index', 'starts', 'expected_goals', 'expected_assists',
       'expected_goal_involvements', 'expected_goals_conceded',
       'Saves_Percentage', 'start_cost', 'end_cost', 'Predicted_End_Cost']]

end_cost_predictions_df_master
end_cost_predictions_df_master.to_csv("C:/Users/Daniel Quinn/Desktop/Bootcamp/Project_2/data/processed/end_cost_predictions_df_master.csv")
end_cost_predictions_df_master.shape
