==> _goalkeeper_analysis/_goalies_analysis/final_work/3-model_goalie_score.ipynb <==
# Here, I am creating a new column in the goalie_hist_model_ready DF that creates a Goalie Score metric for analysis. 

import pandas as pd
import numpy as np
import requests
from sklearn.model_selection import train_test_split
from sklearn.metrics import mean_absolute_error, mean_squared_error, r2_score
from sklearn.linear_model import LinearRegression
from matplotlib import pyplot as plt
from sklearn.ensemble import RandomForestRegressor
from sklearn.preprocessing import StandardScaler


#Import goalie-specific, model ready csv that alrerady has the Saves_percentages and expected total points from the earlier 2 models - this is the master.

goalie_hist_model_ready = pd.read_csv("C:/Users/Daniel Quinn/Desktop/Bootcamp/Project_2/data/processed/total_point_predictions_df_master.csv")
#Ensure we can read the whole dataframe, without "..."
pd.set_option("display.max_rows", None)
goalie_hist_model_ready = goalie_hist_model_ready.drop(columns = ['Unnamed: 0'])
goalie_hist_model_ready.head()


# New calculation for Goalie Score

goalie_hist_model_ready.columns

goalie_score_1 = goalie_hist_model_ready
goalie_score_1["Goalie_Score"]= (goalie_hist_model_ready['end_cost'] + goalie_hist_model_ready['bonus'] + goalie_hist_model_ready['ict_index']) * (goalie_hist_model_ready['Saves_Percentage']/100)
goalie_hist_model_ready.shape


#create X & y variables
X = goalie_hist_model_ready.drop(columns = ['Goalie_Score','id', 'first_name', 'second_name', 'team', 'element_type', 'code',
       'element_code', 'season'])
y = goalie_hist_model_ready['Goalie_Score']

#test-training split
X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=42)

#create model
model_lr = LinearRegression()

model_lr.fit(X_train, y_train)
print("Train model score: ", model_lr.score(X_train, y_train))
print("Test model score: ", model_lr.score(X_test, y_test))

#Make predictions

prediction1 = model_lr.predict(X_test)

#Evaluate models with mse and r2

mse = mean_squared_error(y_test, prediction1) # how close are the predicted values to actual values via the squared differences between expected and real
r2 = r2_score(y_test, prediction1) # r2 - how well do the indep variables explain the variation in the dep var? ) 0 is a perfect model, the larger the nuber, the worse the model is performing

print(f"All Features (Saves_Percentage = y):")
print(f"mean squared error (MSE): {mse}")
print(f"R-squared (R2): {r2}")

#Run the model to get the predicted Goalie Score

predicted_goalie_score = goalie_hist_model_ready.drop(columns = ['Goalie_Score','id', 'first_name', 'second_name', 'team', 'element_type', 'code',
       'element_code', 'season'])
lr_goalie_score = model_lr.predict(predicted_goalie_score)
print(lr_goalie_score)

#Random Forest
#create X & y variables
X = goalie_hist_model_ready.drop(columns=['Goalie_Score','id', 'first_name', 'second_name', 'team', 'element_type', 'code',
       'element_code', 'season'])
y = goalie_hist_model_ready['Goalie_Score']

#test-training split
X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=42)

#create & train model
random_forest = RandomForestRegressor(n_estimators=500, random_state=42).fit(X_train, y_train)

# Evaluate the model
print(f"Training Score: {random_forest.score(X_train, y_train)}")
print(f"Testing Score: {random_forest.score(X_test, y_test)}")

# Feature Importance
feature_importances = random_forest.feature_importances_

feature_importances_df = pd.DataFrame(feature_importances, X.columns)


print(feature_importances_df.sort_values(by=0, ascending=False))

#Run the model to get the predicted Goalie Score from the RFR model

rfr_predicted_goalie_score = goalie_hist_model_ready.drop(columns = ['Goalie_Score','id', 'first_name', 'second_name', 'team', 'element_type', 'code',
       'element_code', 'season'])
rfr_Goalie_Score = random_forest.predict(predicted_goalie_score)
print(rfr_Goalie_Score)

# The Linear Regression model has a 99.67% testing score vs the RFR's of 96.55%, so we will use the LR for the model

# Create new dataframe with expected Goalie Score

goalie_score_prediction_df_master = goalie_hist_model_ready
goalie_score_prediction_df_master['Predicted_Goalie_Score'] = lr_goalie_score
goalie_score_prediction_df_master = goalie_score_prediction_df_master
goalie_score_prediction_df_master.head(15)

goalie_score_prediction_df_master.to_csv("C:/Users/Daniel Quinn/Desktop/Bootcamp/Project_2/data/processed/goalie_score_prediction_df_master.csv")

#total_points_predictions_df_master.to_csv("C:/Users/Daniel Quinn/Desktop/Bootcamp/Project_2/data/processed/total_point_predictions_df_master.csv")
#total_points_predictions_df_master.head(10)

goalie_score_prediction_df_master.shape
