==> _goalkeeper_analysis/_goalies_analysis/final_work/4-Final_Rankings.ipynb <==
# Creating a ranking of goalies from the three models, expected_end_cost, expected_total_points, model_goalie_score

import pandas as pd
import numpy as np
import requests
from sklearn.model_selection import train_test_split
from sklearn.metrics import mean_absolute_error, mean_squared_error, r2_score
from sklearn.linear_model import LinearRegression
from matplotlib import pyplot as plt
from sklearn.ensemble import RandomForestRegressor
from sklearn.preprocessing import StandardScaler


#Import goalie-specific, model ready csv that alrerady has the Saves_percentages and expected total points from the earlier 2 models - this is the master.

final_prediction_df_master = pd.read_csv("C:/Users/Daniel Quinn/Desktop/Bootcamp/Project_2/data/processed/final_prediction_df_master.csv")
#Ensure we can read the whole dataframe, without "..."
pd.set_option("display.max_rows", None)

final_prediction_df_master.columns

#Import team name csv & merge so we have team name

team_name_list_and_id = pd.read_csv("C:/Users/Daniel Quinn/Desktop/Bootcamp/Project_2/data/processed/team_name_list_and_id.csv")
clean_team_name = team_name_list_and_id[['id', 'name', 'short_name']]
clean_team_name_2 = clean_team_name.rename(columns=
                                           {'id': 'team', 'name': 'team_name'})
clean_team_name_2 = clean_team_name_2[['team', 'team_name']]
clean_team_name_2

final_ranking = pd.merge(clean_team_name_2, final_prediction_df_master, on="team", how="inner")
final_ranking_dropna = final_ranking.dropna()
final_ranking_dropna.drop(columns=['Unnamed: 0'])
final_goalie_rank = final_ranking_dropna.sort_values('Final_Point_Ranking', ascending=False)
final_goalie_rank.columns

final_goalie_rank = final_goalie_rank[
    ['id', 'first_name', 'second_name',
       'element_type', 'code', 'element_code', 'Saves_Percentage', 'Predicted_End_Cost',
       'Predicted_Total_Points', 'Goalie_Score',
       'Predicted_Goalie_Score', 'Composite Score', 'Final_Point_Ranking']]

final_goalie_rank.to_csv('C:/Users/Daniel Quinn/Desktop/Bootcamp/Project_2/data/processed/final_goalie_rank.csv')
final_goalie_rank.head()
