==> _goalkeeper_analysis/_goalies_analysis/final_work/5- for_merge_in_model_FINAL.ipynb <==
# Using the total_points dataset to determine goalie rankings

import pandas as pd
import numpy as np
import requests
from sklearn.model_selection import train_test_split
from sklearn.metrics import mean_absolute_error, mean_squared_error, r2_score
from sklearn.linear_model import LinearRegression
from matplotlib import pyplot as plt
from sklearn.ensemble import RandomForestRegressor
from sklearn.preprocessing import StandardScaler


#Import goalie-specific, model ready csv

all_goalies = pd.read_csv("C:/Users/Daniel Quinn/Desktop/Bootcamp/Project_2/data/processed/end_cost_predictions_df_master.csv")

#Ensure we can read the whole dataframe, without "..."
pd.set_option("display.max_rows", None)
all_goalies = all_goalies.drop(columns = ['Unnamed: 0'])
all_goalies.head()


#Create Saves Percentage stat to screen irrelevant players via a dropna():
all_goalies_dropna = all_goalies
all_goalies_dropna["Saves_Percentage"] = (all_goalies['saves'] + all_goalies['penalties_saved']) / (all_goalies['saves'] + all_goalies['penalties_saved'] + all_goalies['goals_conceded']) * 100
#pd.set_option("display.max_columns", None)
all_goalies_dropna = all_goalies.dropna()
all_goalies_dropna.shape


#Create a better dataframe name for going forward

#goalie_hist_model_2_ready 
goalie_hist_model_1 = all_goalies_dropna
goalie_hist_model_1.columns

# Regressions for total_points

#Linear Regression
# #create X & y variables
X = goalie_hist_model_1[['minutes', 'goals_scored', 'clean_sheets', 'penalties_saved', 
                              'penalties_missed', 'bps', 'bonus', 'yellow_cards',
                              'red_cards', 'own_goals', 'goals_conceded']]
y = goalie_hist_model_1['total_points']

#test-training split
X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=42)

#create model
lr_model = LinearRegression()

lr_model.fit(X_train, y_train)
print("Train model score: ", lr_model.score(X_train, y_train))
print("Test model score: ", lr_model.score(X_test, y_test))

#Make predictions

prediction1 = lr_model.predict(X_test)

#Evaluate models with mse and r2

mse = mean_squared_error(y_test, prediction1) # how close are the predicted values to actual values via the squared differences between expected and real
r2 = r2_score(y_test, prediction1) # r2 - how well do the indep variables explain the variation in the dep var? ) 0 is a perfect model, the larger the nuber, the worse the model is performing

print(f"All Features (Saves_Percentage = y):")
print(f"mean squared error (MSE): {mse}")
print(f"R-squared (R2): {r2}")

# RFR
I deleted "minutes" from the below RFR model because it was given way too much importance. With it gone, "saves" moves up dramatically (which makes sense, given that these are goalies...) Doing so made almost no difference in the Linear Regression above.

#Random Forest
# #create X & y variables
X = goalie_hist_model_1[['minutes', 'goals_scored', 'clean_sheets', 'penalties_saved', 
                              'penalties_missed', 'bps', 'bonus', 'yellow_cards',
                              'red_cards', 'own_goals', 'goals_conceded']]
y = goalie_hist_model_1['total_points']

#test-training split
X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=42)

#create & train model
random_forest = RandomForestRegressor(n_estimators=500, random_state=42).fit(X_train, y_train)

# Evaluate the model
print(f"Training Score: {random_forest.score(X_train, y_train)}")
print(f"Testing Score: {random_forest.score(X_test, y_test)}")

# Feature Importance
feature_importances = random_forest.feature_importances_

feature_importances_df = pd.DataFrame(feature_importances, X.columns)


print(feature_importances_df.sort_values(by=0, ascending=False))



# Predictions:
There is no difference between the LR & RFR predicted values:

#LR Predict:

predicted_total_points_lr = goalie_hist_model_1[['minutes', 'goals_scored', 'clean_sheets', 'penalties_saved', 
                              'penalties_missed', 'bps', 'bonus', 'yellow_cards',
                              'red_cards', 'own_goals', 'goals_conceded']]

y_new_pred_total_points_lr = lr_model.predict(predicted_total_points_lr)
y_new_pred_total_points_lr

exp_total_points_lr = pd.DataFrame(y_new_pred_total_points_lr)
exp_total_points_lr.head(15)



#RFR Predict:

predicted_total_points_rfr = goalie_hist_model_1[['minutes', 'goals_scored', 'clean_sheets', 'penalties_saved', 
                              'penalties_missed', 'bps', 'bonus', 'yellow_cards',
                              'red_cards', 'own_goals', 'goals_conceded']]

y_new_pred_total_points_rfr = random_forest.predict(predicted_total_points_rfr)
y_new_pred_total_points_rfr

exp_total_points_rfr = pd.DataFrame(y_new_pred_total_points_rfr)
exp_total_points_rfr.head(15)


# Create new dataframe with expected total_points

total_points_predictions_df_master = goalie_hist_model_1
total_points_predictions_df_master['Predicted_Total_Points'] = exp_total_points_rfr
total_points_predictions_df_master = total_points_predictions_df_master[['id', 'first_name', 'second_name', 'team', 'element_type', 'code',
       'element_code', 'season', 'minutes', 'total_points', 'Predicted_Total_Points']]

total_points_predictions_df_master.head(10)

total_points_predictions_df_master.to_csv("C:/Users/Daniel Quinn/Desktop/Bootcamp/Project_2/data/processed/total_point_predictions_df_master.csv")
total_points_predictions_df_master.sort_values('second_name', ascending=True)
total_points_predictions_df_master.columns


# Create new dataframe with expected total_points

total_points_predictions_df_master = goalie_hist_model_1
total_points_predictions_df_master['Predicted_Total_Points'] = exp_total_points_rfr
total_points_predictions_df_master = total_points_predictions_df_master[['id', 'first_name', 'second_name', 'team', 'element_type', 'code',
       'element_code', 'season', 'minutes', 'total_points', 'Predicted_Total_Points']]

total_points_predictions_df_master.head(10)

#export to csv
total_points_predictions_df_master.to_csv("C:/Users/Daniel Quinn/Desktop/Bootcamp/Project_2/data/processed/total_point_predictions_df_master.csv")
total_points_predictions_df_master.sort_values('second_name', ascending=True)

#edit dataframe down for final merge
goalie_final_merge = total_points_predictions_df_master[['id', 'season', 'first_name', 'second_name', 'Predicted_Total_Points']]
goalie_final_merge = goalie_final_merge[goalie_final_merge['season'] == 2023]
goalie_final_merge.sort_values('Predicted_Total_Points', ascending=False)

#scale data for congruence with Matt's dataset from Sorare
goalie_final_merge["Predicted_Total_Points"] = goalie_final_merge["Predicted_Total_Points"]/2

goalie_final_merge.head()

#export final document for merge into model
goalie_final_merge.to_csv('C:/Users/Daniel Quinn/Desktop/Bootcamp/Project_2/data/processed/goalie_final_merge.csv')
